fix:Error
エラー直した

+++ Assets/Scripts/Enemy/Boss/Missile/SC_FallingMissile.cs
@@ -94,7 +94,7 @@ public class SC_FallingMissile : MonoBehaviour, SC_IPoolObject
 
         if (playerHP != null)
         {
-            playerHP.TakeDamage(playerDamage);
+            playerHP.TakeDamage((int)playerDamage);
         }
 
         if (other.CompareTag("Player"))
