+++ Assets/Scripts/Enemy/Boss/Missile/SC_HomingMissile.cs
@@ -164,7 +164,7 @@ public class SC_HomingMissile : MonoBehaviour, SC_IPoolObject
 
         if (playerHP != null)
         {
-            playerHP.TakeDamage(playerDamage);
+            playerHP.TakeDamage((int)playerDamage);
         }
 
         if (other.CompareTag("Player"))
