+++ Assets/Scripts/Enemy/Boss/Missile/SC_ReflectableMissile.cs
@@ -145,7 +145,7 @@ public class SC_ReflectableMissile : MonoBehaviour, SC_IPoolObject
 
                 if (playerHP != null)
                 {
-                    playerHP.TakeDamage(playerDamage);
+                    playerHP.TakeDamage((int)playerDamage);
                 }
 
                 ReturnToPool();
