+++ Assets/Scripts/Enemy/Boss/Missile/SC_StraightMissile.cs
@@ -72,7 +72,7 @@ public class SC_StraightMissile : MonoBehaviour, SC_IPoolObject
 
         if (playerHP != null)
         {
-            playerHP.TakeDamage(playerDamage);
+            playerHP.TakeDamage((int)playerDamage);
         }
 
         if (other.CompareTag("Player"))
